requirements.txt
crewai
crewai-tools
litellm
google-genai
requests
python-dotenv
streamlit